dashboard-mvp/vk-python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vk-ads-pipeline"
version = "0.2.0"
description = "Сервис синхронизации метрик VK Ads с Google Sheets."
authors = [{ name = "Zakaz Analytics Team" }]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27,<0.28",
    "python-dotenv>=1.0,<2.0",
    "google-api-python-client>=2.148",
    "google-auth>=2.35",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "freezegun>=1.5,<2.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = "-q"
